+++ components/deeplink/impl/src/main/java/com/flipperdevices/deeplink/impl/parser/delegates/DeepLinkFileUriGrantPermission.kt
@@ -9,6 +9,8 @@ import com.flipperdevices.deeplink.api.DeepLinkParserDelegate
 import com.flipperdevices.deeplink.impl.parser.filename
 import com.flipperdevices.deeplink.impl.parser.length
 import com.flipperdevices.deeplink.model.DeepLinkParserDelegatePriority
+import com.flipperdevices.core.ktx.jre.filename
+import com.flipperdevices.core.ktx.jre.length
 import com.flipperdevices.deeplink.model.Deeplink
 import com.flipperdevices.deeplink.model.DeeplinkContent
 import com.squareup.anvil.annotations.ContributesMultibinding