+++ components/updater/downloader/src/main/java/com/flipperdevices/updater/downloader/api/DownloadAndUnpackDelegateApiImpl.kt
@@ -3,6 +3,7 @@ package com.flipperdevices.updater.downloader.api
 import com.flipperdevices.core.di.AppGraph
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.error
+import com.flipperdevices.updater.api.DownloadAndUnpackDelegateApi
 import com.flipperdevices.updater.model.DistributionFile
 import com.squareup.anvil.annotations.ContributesBinding
 import io.ktor.client.HttpClient
@@ -23,20 +24,10 @@ import org.apache.commons.compress.compressors.gzip.GzipCompressorInputStream
 
 private const val TRY_MAX_COUNT = 3
 
-interface DownloadAndUnpackDelegate {
-    suspend fun download(
-        distributionFile: DistributionFile,
-        target: File,
-        onProgress: (suspend (Long, Long) -> Unit)? = null
-    )
-
-    suspend fun unpack(source: File, target: File)
-}
-
-@ContributesBinding(AppGraph::class, DownloadAndUnpackDelegate::class)
-class DownloadAndUnpackDelegateImpl @Inject constructor(
+@ContributesBinding(AppGraph::class, DownloadAndUnpackDelegateApi::class)
+class DownloadAndUnpackDelegateApiImpl @Inject constructor(
     private val client: HttpClient
-) : DownloadAndUnpackDelegate, LogTagProvider {
+) : DownloadAndUnpackDelegateApi, LogTagProvider {
     override val TAG = "DownloadAndUnpackDelegate"
 
     override suspend fun download(
@@ -79,7 +70,7 @@ class DownloadAndUnpackDelegateImpl @Inject constructor(
             String(Hex.encodeHex(DigestUtils.sha256(it)))
         }
 
-        if (hash != distributionFile.sha256) {
+        if (hash != distributionFile.sha256 && distributionFile.sha256 != null) {
             throw IllegalStateException(
                 "Hash mismatch. Expected: ${distributionFile.sha256}, actual: $hash"
             )